requirements.txt
Django==2.0.1
djangorestframework==3.7.7
django-cors-headers==2.1.0
mysqlclient==1.3.12